refactor: create pickReport function to avoid duplication
