+++ lib/prepare_security.js
@@ -1,4 +1,3 @@
-import nv from '@pkgjs/nv';
 import fs from 'node:fs';
 import path from 'node:path';
 import auth from './auth.js';
@@ -10,10 +9,10 @@ import {
   PLACEHOLDERS,
   checkoutOnSecurityReleaseBranch,
   commitAndPushVulnerabilitiesJSON,
-  getSummary,
   validateDate,
   promptDependencies,
-  getSupportedVersions
+  getSupportedVersions,
+  pickReport
 } from './security-release/security-release.js';
 import _ from 'lodash';
 
@@ -152,48 +151,12 @@ export default class PrepareSecurityRelease {
   async chooseReports() {
     this.cli.info('Getting triaged H1 reports...');
     const reports = await this.req.getTriagedReports();
-    const supportedVersions = (await nv('supported'))
-      .map((v) => `${v.versionName}.x`)
-      .join(',');
     const selectedReports = [];
 
     for (const report of reports.data) {
-      const {
-        id, attributes: { title, cve_ids },
-        relationships: { severity, weakness, reporter }
-      } = report;
-      const link = `https://hackerone.com/reports/${id}`;
-      const reportSeverity = {
-        rating: severity?.data?.attributes?.rating || '',
-        cvss_vector_string: severity?.data?.attributes?.cvss_vector_string || '',
-        weakness_id: weakness?.data?.id || ''
-      };
-
-      this.cli.separator();
-      this.cli.info(`Report: ${link} - ${title} (${reportSeverity?.rating})`);
-      const include = await this.cli.prompt(
-        'Would you like to include this report to the next security release?',
-        { defaultAnswer: true });
-      if (!include) {
-        continue;
-      }
-
-      const versions = await this.cli.prompt('Which active release lines this report affects?', {
-        questionType: 'input',
-        defaultAnswer: supportedVersions
-      });
-      const summaryContent = await getSummary(id, this.req);
-
-      selectedReports.push({
-        id,
-        title,
-        cveIds: cve_ids,
-        severity: reportSeverity,
-        summary: summaryContent ?? '',
-        affectedVersions: versions.split(',').map((v) => v.replace('v', '').trim()),
-        link,
-        reporter: reporter.data.attributes.username
-      });
+      const rep = await pickReport(report);
+      if (!rep) continue;
+      selectedReports.push(rep);
     }
     return selectedReports;
   }
