+++ lib/security-announcement.js
@@ -40,8 +40,10 @@ export default class SecurityAnnouncement {
     validateDate(content.releaseDate);
     const releaseDate = new Date(content.releaseDate);
 
-    await Promise.all([this.createDockerNodeIssue(releaseDate),
-      this.createBuildWGIssue(releaseDate)]);
+    await Promise.all([
+      this.createDockerNodeIssue(releaseDate),
+      this.createBuildWGIssue(releaseDate)
+    ]);
   }
 
   async createBuildWGIssue(releaseDate) {
@@ -57,8 +59,8 @@ export default class SecurityAnnouncement {
   createPreleaseAnnouncementIssue(releaseDate, team) {
     const title = `[NEXT-SECURITY-RELEASE] Heads up on upcoming Node.js\
  security release ${formatDateToYYYYMMDD(releaseDate)}`;
-    const content = 'As per security release workflow,' +
-      ` creating issue to give the ${team} team a heads up.`;
+    const content = `As per security release workflow,\
+ creating issue to give the ${team} team a heads up.`;
     return { title, content };
   }
