+++ lib/security-release/security-release.js
@@ -124,3 +124,42 @@ export async function createIssue(title, content, repository, { cli, req }) {
     process.exit(1);
   }
 }
+
+export async function pickReport(report) {
+  const {
+    id, attributes: { title, cve_ids },
+    relationships: { severity, weakness, reporter }
+  } = report;
+  const link = `https://hackerone.com/reports/${id}`;
+  const reportSeverity = {
+    rating: severity?.data?.attributes?.rating || '',
+    cvss_vector_string: severity?.data?.attributes?.cvss_vector_string || '',
+    weakness_id: weakness?.data?.id || ''
+  };
+
+  this.cli.separator();
+  this.cli.info(`Report: ${link} - ${title} (${reportSeverity?.rating})`);
+  const include = await this.cli.prompt(
+    'Would you like to include this report to the next security release?',
+    { defaultAnswer: true });
+  if (!include) {
+    return;
+  }
+
+  const versions = await this.cli.prompt('Which active release lines this report affects?', {
+    questionType: 'input',
+    defaultAnswer: await getSupportedVersions()
+  });
+  const summaryContent = await getSummary(id, this.req);
+
+  return {
+    id,
+    title,
+    cveIds: cve_ids,
+    severity: reportSeverity,
+    summary: summaryContent ?? '',
+    affectedVersions: versions.split(',').map((v) => v.replace('v', '').trim()),
+    link,
+    reporter: reporter.data.attributes.username
+  };
+}
