+++ lib/update_security_release.js
@@ -3,9 +3,8 @@ import {
   NEXT_SECURITY_RELEASE_REPOSITORY,
   checkoutOnSecurityReleaseBranch,
   commitAndPushVulnerabilitiesJSON,
-  getSupportedVersions,
-  getSummary,
-  validateDate
+  validateDate,
+  pickReport
 } from './security-release/security-release.js';
 import fs from 'node:fs';
 import path from 'node:path';
@@ -81,44 +80,14 @@ export default class UpdateSecurityRelease {
 
     // get h1 report
     const { data: report } = await req.getReport(reportId);
-    const {
-      id, attributes: { title, cve_ids },
-      relationships: { severity, reporter, weakness }
-    } = report;
-
-    const reportSeverity = {
-      rating: severity?.data?.attributes?.rating || '',
-      cvss_vector_string: severity?.data?.attributes?.cvss_vector_string || '',
-      weakness_id: weakness?.data?.id || ''
-    };
-
-    // get the affected versions
-    const supportedVersions = await getSupportedVersions();
-    const versions = await cli.prompt('Which active release lines this report affects?', {
-      questionType: 'input',
-      defaultAnswer: supportedVersions
-    });
-
-    // get the team summary from h1 report
-    const summaryContent = await getSummary(id, req);
-
-    const entry = {
-      id,
-      title,
-      link: `https://hackerone.com/reports/${id}`,
-      cveIds: cve_ids,
-      severity: reportSeverity,
-      summary: summaryContent ?? '',
-      affectedVersions: versions.split(',').map((v) => v.replace('v', '').trim()),
-      reporter: reporter.data.attributes.username
-    };
+    const entry = await pickReport(report);
 
     const vulnerabilitiesJSONPath = this.getVulnerabilitiesJSONPath();
     const content = this.readVulnerabilitiesJSON(vulnerabilitiesJSONPath);
     content.reports.push(entry);
     fs.writeFileSync(vulnerabilitiesJSONPath, JSON.stringify(content, null, 2));
-    this.cli.ok(`Updated vulnerabilities.json with the report: ${id}`);
-    const commitMessage = `chore: added report ${id} to vulnerabilities.json`;
+    this.cli.ok(`Updated vulnerabilities.json with the report: ${entry.id}`);
+    const commitMessage = `chore: added report ${entry.id} to vulnerabilities.json`;
     commitAndPushVulnerabilitiesJSON(vulnerabilitiesJSONPath,
       commitMessage, { cli, repository: this.repository });
     cli.ok('Done!');
